Use FileScoped namespaces and refactor tests.
