+++ samples/complexapp/complexapp/Program.cs
@@ -1,19 +1,17 @@
-﻿using System;
-using libbar;
+﻿using libbar;
 using libfoo;
 using static System.Console;
 
-namespace complexapp
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
-            var initialString = StringLibrary.GetString();
-            var reversedString = StringUtils.ReverseString(initialString);
+namespace complexapp;
 
-            WriteLine($"string: {initialString}");
-            WriteLine($"reversed: {reversedString}");
-        }
+class Program
+{
+    static void Main(string[] args)
+    {
+        string initialString = StringLibrary.GetString();
+        string reversedString = StringUtils.ReverseString(initialString);
+
+        WriteLine($"string: {initialString}");
+        WriteLine($"reversed: {reversedString}");
     }
 }
