+++ samples/complexapp/libbar/Bar.cs
@@ -1,15 +1,16 @@
 ﻿using System;
 
-namespace libbar
+namespace libbar;
+
+public class StringUtils
 {
-    public class StringUtils
+    public static string ReverseString(string input)
     {
-        public static string ReverseString(string input)
-        {
-            var chars = input.ToCharArray();
-            Array.Reverse(chars);
-            var reversedString = new string(chars);
-            return reversedString;
-        }
+        ArgumentNullException.ThrowIfNull(input);
+
+        var chars = input.ToCharArray();
+        Array.Reverse(chars);
+        var reversedString = new string(chars);
+        return reversedString;
     }
 }
