+++ samples/complexapp/libfoo/Foo.cs
@@ -1,9 +1,6 @@
-﻿using System;
+﻿namespace libfoo;
 
-namespace libfoo
+public class StringLibrary
 {
-    public class StringLibrary
-    {
-       public static string GetString() => "The quick brown fox jumps over the lazy dog"; 
-    }
+    public static string GetString() => "The quick brown fox jumps over the lazy dog";
 }
